backend/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-aiplatform
python-dotenv
mangum
